Demo of using HTTP/2 prior knowledge over TLS
There are a couple things that make us less excited about this
implementation:
1. We need to use http2.Transport instead of http.Transport, which is
less used now that http.Transport supports HTTP/2
2. The http2.Transport doesn't support a DialTLSContext hook, so we lose
the context when dialing. This is not recommended for Golang 1.15+.

In addition to the context, we may be losing some other behavior when
switching to the HTTP/2 transport and overriding the DialTLS function.
It would be better to use ALPN over TLS, whenever possible.

Co-authored-by: Maria Shaldybin <mariash@vmware.com>